server/requirements.txt
celery==5.3.6
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.5
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psutil==5.9.6
drf-extensions
bitstring==3.1.9
chardet==4.0.0
jsonpath==0.82
jsonpath_rw==1.4.0
tabulate==0.8.9
websocket_client==1.3.2
wechatpayv3
